+++ Magna/src/main/java/de/morigm/magna/api/helper/TranslationHelper.java
@@ -1,13 +1,13 @@
 package de.morigm.magna.api.helper;
 
-import de.morigm.magna.Main;
+import de.morigm.magna.api.Magna;
 
 public interface TranslationHelper
 {
 
 	public default String translate(String text)
 	{
-		return Main.getInstance().getLanguage().translate(text);
+		return Magna.getLanguage().translate(text);
 	}
 	
 }